gen_api/requirements.txt
accelerate
diffusers
edge_tts
Flask
moviepy
numpy
opencv-python
montreal-forced-aligner
pedalboard
pydub
torch
transformers
waitress
pytest
pytest-asyncio
pytest-cov
setuptools<60.0
bs4
lxml
aeneas
python-dotenv